Motivational Quotes
Code chooses a random quote from test.txt file, prints that quote. Counts the words and characters used in the quote and prints them too. I've left a lot of instructions in the code as comments because the code will breake as soon as you don't follow the same quotation format i've used.