Medium
Yes, this simple question was a medium on LeetCode. It is medium because it was noted that answer must not use built-in BigInteger library or convert the inputs to integer directly. However both of the solutions they provided already used "multiplication = int(digit1) * int(digit2) + carry". So if their solution is acceptable so is mine.